Move selectReport tests to builder package.
