+++ execute/plugin_test.go
@@ -3,7 +3,6 @@ package execute
 import (
 	"context"
 	"fmt"
-	"math/rand"
 	"testing"
 	"time"
 
@@ -19,65 +18,17 @@ import (
 	"github.com/smartcontractkit/libocr/offchainreporting2plus/types"
 	libocrtypes "github.com/smartcontractkit/libocr/ragep2p/types"
 
-	"github.com/smartcontractkit/chainlink-common/pkg/hashutil"
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
-	"github.com/smartcontractkit/chainlink-common/pkg/merklemulti"
 	cciptypes "github.com/smartcontractkit/chainlink-common/pkg/types/ccipocr3"
 
-	"github.com/smartcontractkit/chainlink-ccip/internal/plugincommon"
-
-	"github.com/smartcontractkit/chainlink-ccip/execute/report"
 	"github.com/smartcontractkit/chainlink-ccip/internal/libs/slicelib"
 	"github.com/smartcontractkit/chainlink-ccip/internal/mocks"
+	"github.com/smartcontractkit/chainlink-ccip/internal/plugincommon"
 	"github.com/smartcontractkit/chainlink-ccip/internal/reader"
 	reader_mock "github.com/smartcontractkit/chainlink-ccip/internal/reader/mocks"
 	"github.com/smartcontractkit/chainlink-ccip/pluginconfig"
 	"github.com/smartcontractkit/chainlink-ccip/plugintypes"
 )
-
-// makeMessage creates a message deterministically derived from the given inputs.
-func makeMessage(src cciptypes.ChainSelector, num cciptypes.SeqNum, nonce uint64) cciptypes.Message {
-	var placeholderID cciptypes.Bytes32
-	n, err := rand.New(rand.NewSource(int64(src) * int64(num) * int64(nonce))).Read(placeholderID[:])
-	if n != 32 {
-		panic(fmt.Sprintf("Unexpected number of bytes read for placeholder id: want 32, got %d", n))
-	}
-	if err != nil {
-		panic(fmt.Sprintf("Error reading random bytes: %v", err))
-	}
-
-	return cciptypes.Message{
-		Header: cciptypes.RampMessageHeader{
-			MessageID:           placeholderID,
-			SourceChainSelector: src,
-			SequenceNumber:      num,
-			MsgHash:             cciptypes.Bytes32{},
-			Nonce:               nonce,
-		},
-	}
-}
-
-// mustParseByteStr parses a given string into a byte array, any error causes a panic. Pass in an empty string for a
-// random byte array.
-// nolint:unparam // surly this will be useful at some point...
-func mustParseByteStr(byteStr string) cciptypes.Bytes32 {
-	if byteStr == "" {
-		var randomBytes cciptypes.Bytes32
-		n, err := rand.New(rand.NewSource(0)).Read(randomBytes[:])
-		if n != 32 {
-			panic(fmt.Sprintf("Unexpected number of bytes read for placeholder id: want 32, got %d", n))
-		}
-		if err != nil {
-			panic(fmt.Sprintf("Error reading random bytes: %v", err))
-		}
-		return randomBytes
-	}
-	b, err := cciptypes.NewBytes32FromString(byteStr)
-	if err != nil {
-		panic(err)
-	}
-	return b
-}
 
 func Test_getPendingExecutedReports(t *testing.T) {
 	tests := []struct {
@@ -213,344 +164,6 @@ func Test_getPendingExecutedReports(t *testing.T) {
 	}
 }
 
-// TODO: better than this
-type tdr struct{}
-
-func (t tdr) ReadTokenData(
-	ctx context.Context, srcChain cciptypes.ChainSelector, num cciptypes.SeqNum) ([][]byte, error,
-) {
-	return nil, nil
-}
-
-// breakCommitReport by adding an extra message. This causes the report to have an unexpected number of messages.
-func breakCommitReport(
-	commitReport plugintypes.ExecutePluginCommitDataWithMessages,
-) plugintypes.ExecutePluginCommitDataWithMessages {
-	commitReport.Messages = append(commitReport.Messages, cciptypes.Message{})
-	return commitReport
-}
-
-// makeTestCommitReport creates a basic commit report with messages given different parameters. This function
-// will panic if the input parameters are inconsistent.
-func makeTestCommitReport(
-	hasher cciptypes.MessageHasher,
-	numMessages,
-	srcChain,
-	firstSeqNum,
-	block int,
-	timestamp int64,
-	rootOverride cciptypes.Bytes32,
-	executed []cciptypes.SeqNum,
-) plugintypes.ExecutePluginCommitDataWithMessages {
-	sequenceNumberRange :=
-		cciptypes.NewSeqNumRange(cciptypes.SeqNum(firstSeqNum), cciptypes.SeqNum(firstSeqNum+numMessages-1))
-
-	for _, e := range executed {
-		if !sequenceNumberRange.Contains(e) {
-			panic("executed message out of range")
-		}
-	}
-	var messages []cciptypes.Message
-	for i := 0; i < numMessages; i++ {
-		messages = append(messages, makeMessage(
-			cciptypes.ChainSelector(srcChain),
-			cciptypes.SeqNum(i+firstSeqNum),
-			uint64(i)))
-	}
-
-	commitReport := plugintypes.ExecutePluginCommitDataWithMessages{
-		ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
-			// MerkleRoot:          root,
-			SourceChain:         cciptypes.ChainSelector(srcChain),
-			SequenceNumberRange: sequenceNumberRange,
-			Timestamp:           time.UnixMilli(timestamp),
-			BlockNum:            uint64(block),
-			ExecutedMessages:    executed,
-		},
-		Messages: messages,
-	}
-
-	// calculate merkle root
-	root := rootOverride
-	if root.IsEmpty() {
-		tree, err := report.ConstructMerkleTree(context.Background(), hasher, commitReport)
-		if err != nil {
-			panic(fmt.Sprintf("unable to construct merkle tree: %s", err))
-		}
-		commitReport.MerkleRoot = tree.Root()
-	}
-
-	return commitReport
-}
-
-// assertMerkleRoot computes the source messages merkle root, then computes a verification with the proof, then compares
-// the roots.
-func assertMerkleRoot(
-	t *testing.T,
-	hasher cciptypes.MessageHasher,
-	execReport cciptypes.ExecutePluginReportSingleChain,
-	commitReport plugintypes.ExecutePluginCommitDataWithMessages,
-) {
-	keccak := hashutil.NewKeccak()
-	// Generate merkle root from commit report messages
-	var leafHashes [][32]byte
-	for _, msg := range commitReport.Messages {
-		hash, err := hasher.Hash(context.Background(), msg)
-		require.NoError(t, err)
-		leafHashes = append(leafHashes, hash)
-	}
-	tree, err := merklemulti.NewTree(keccak, leafHashes)
-	require.NoError(t, err)
-	merkleRoot := tree.Root()
-
-	// Generate merkle root from exec report messages and proof
-	ctx := context.Background()
-	var leaves [][32]byte
-	for _, msg := range execReport.Messages {
-		hash, err := hasher.Hash(ctx, msg)
-		require.NoError(t, err)
-		leaves = append(leaves, hash)
-	}
-	proofCast := make([][32]byte, len(execReport.Proofs))
-	for i, p := range execReport.Proofs {
-		copy(proofCast[i][:], p[:32])
-	}
-	var proof merklemulti.Proof[[32]byte]
-	proof.Hashes = proofCast
-	proof.SourceFlags = slicelib.BitFlagsToBools(execReport.ProofFlagBits.Int, len(leaves)+len(proofCast)-1)
-	recomputedMerkleRoot, err := merklemulti.VerifyComputeRoot(hashutil.NewKeccak(),
-		leaves,
-		proof)
-	assert.NoError(t, err)
-	assert.NotNil(t, recomputedMerkleRoot)
-
-	// Compare them
-	assert.Equal(t, merkleRoot, recomputedMerkleRoot)
-}
-
-func Test_selectReport(t *testing.T) {
-	hasher := mocks.NewMessageHasher()
-	codec := mocks.NewExecutePluginJSONReportCodec()
-	lggr := logger.Test(t)
-	var tokenDataReader tdr
-
-	type args struct {
-		reports       []plugintypes.ExecutePluginCommitDataWithMessages
-		maxReportSize int
-	}
-	tests := []struct {
-		name                  string
-		args                  args
-		expectedExecReports   int
-		expectedCommitReports int
-		expectedExecThings    []int
-		lastReportExecuted    []cciptypes.SeqNum
-		wantErr               string
-	}{
-		{
-			name: "empty report",
-			args: args{
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{},
-			},
-			expectedExecReports:   0,
-			expectedCommitReports: 0,
-		},
-		{
-			name: "half report",
-			args: args{
-				maxReportSize: 2300,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						nil),
-				},
-			},
-			expectedExecReports:   1,
-			expectedCommitReports: 1,
-			expectedExecThings:    []int{5},
-			lastReportExecuted:    []cciptypes.SeqNum{100, 101, 102, 103, 104},
-		},
-		{
-			name: "full report",
-			args: args{
-				maxReportSize: 10000,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						nil),
-				},
-			},
-			expectedExecReports:   1,
-			expectedCommitReports: 0,
-			expectedExecThings:    []int{10},
-		},
-		{
-			name: "two reports",
-			args: args{
-				maxReportSize: 15000,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						nil),
-					makeTestCommitReport(hasher, 20, 2, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						nil),
-				},
-			},
-			expectedExecReports:   2,
-			expectedCommitReports: 0,
-			expectedExecThings:    []int{10, 20},
-		},
-		{
-			name: "one and half reports",
-			args: args{
-				maxReportSize: 8500,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						nil),
-					makeTestCommitReport(hasher, 20, 2, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						nil),
-				},
-			},
-			expectedExecReports:   2,
-			expectedCommitReports: 1,
-			expectedExecThings:    []int{10, 10},
-			lastReportExecuted:    []cciptypes.SeqNum{100, 101, 102, 103, 104, 105, 106, 107, 108, 109},
-		},
-		{
-			name: "exactly one report",
-			args: args{
-				maxReportSize: 4200,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						nil),
-					makeTestCommitReport(hasher, 20, 2, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						nil),
-				},
-			},
-			expectedExecReports:   1,
-			expectedCommitReports: 1,
-			expectedExecThings:    []int{10},
-			lastReportExecuted:    []cciptypes.SeqNum{},
-		},
-		{
-			name: "execute remainder of partially executed report",
-			args: args{
-				maxReportSize: 2500,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						[]cciptypes.SeqNum{100, 101, 102, 103, 104}),
-				},
-			},
-			expectedExecReports:   1,
-			expectedCommitReports: 0,
-			expectedExecThings:    []int{5},
-		},
-		{
-			name: "partially execute remainder of partially executed report",
-			args: args{
-				maxReportSize: 2050,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						[]cciptypes.SeqNum{100, 101, 102, 103, 104}),
-				},
-			},
-			expectedExecReports:   1,
-			expectedCommitReports: 1,
-			expectedExecThings:    []int{4},
-			lastReportExecuted:    []cciptypes.SeqNum{100, 101, 102, 103, 104, 105, 106, 107, 108},
-		},
-		{
-			name: "execute remainder of sparsely executed report",
-			args: args{
-				maxReportSize: 3500,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						[]cciptypes.SeqNum{100, 102, 104, 106, 108}),
-				},
-			},
-			expectedExecReports:   1,
-			expectedCommitReports: 0,
-			expectedExecThings:    []int{5},
-		},
-		{
-			name: "partially execute remainder of partially executed sparse report",
-			args: args{
-				maxReportSize: 2050,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						cciptypes.Bytes32{}, // generate a correct root.
-						[]cciptypes.SeqNum{100, 102, 104, 106, 108}),
-				},
-			},
-			expectedExecReports:   1,
-			expectedCommitReports: 1,
-			expectedExecThings:    []int{4},
-			lastReportExecuted:    []cciptypes.SeqNum{100, 101, 102, 103, 104, 105, 106, 107, 108},
-		},
-		{
-			name: "broken report",
-			args: args{
-				maxReportSize: 10000,
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					breakCommitReport(makeTestCommitReport(hasher, 10, 1, 101, 1000, 10101010102,
-						cciptypes.Bytes32{}, // generate a correct root.
-						nil)),
-				},
-			},
-			wantErr: "unable to build a single chain report",
-		},
-		{
-			name: "invalid merkle root",
-			args: args{
-				reports: []plugintypes.ExecutePluginCommitDataWithMessages{
-					makeTestCommitReport(hasher, 10, 1, 100, 999, 10101010101,
-						mustParseByteStr(""), // random root
-						nil),
-				},
-			},
-			wantErr: "merkle root mismatch: expected 0x00000000000000000",
-		},
-		// TODO: A test that requires skipping over a large message because only a smaller message fits in the report.
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			ctx := context.Background()
-			execReports, commitReports, err :=
-				selectReport(ctx, lggr, hasher, codec, tokenDataReader, tt.args.reports, tt.args.maxReportSize)
-			if tt.wantErr != "" {
-				assert.Contains(t, err.Error(), tt.wantErr)
-				return
-			}
-			require.NoError(t, err)
-			require.Len(t, execReports, tt.expectedExecReports)
-			require.Len(t, commitReports, tt.expectedCommitReports)
-			for i, execReport := range execReports {
-				require.Lenf(t, execReport.Messages, tt.expectedExecThings[i],
-					"Unexpected number of messages, iter %d", i)
-				require.Lenf(t, execReport.OffchainTokenData, tt.expectedExecThings[i],
-					"Unexpected number of token data, iter %d", i)
-				require.NotEmptyf(t, execReport.Proofs, "Proof should not be empty.")
-				assertMerkleRoot(t, hasher, execReport, tt.args.reports[i])
-			}
-			// If the last report is partially executed, the executed messages can be checked.
-			if len(execReports) > 0 && len(tt.lastReportExecuted) > 0 {
-				lastReport := commitReports[len(commitReports)-1]
-				require.ElementsMatch(t, tt.lastReportExecuted, lastReport.ExecutedMessages)
-			}
-		})
-	}
-}
-
 func TestPlugin_Close(t *testing.T) {
 	mockReader := mocks.NewCCIPReader()
 	mockReader.On("Close", mock.Anything).Return(nil)
@@ -642,7 +255,7 @@ func TestPlugin_ValidateObservation_ValidateObservedSeqNum_Error(t *testing.T) {
 	}
 
 	// Reports with duplicate roots.
-	root := mustParseByteStr("")
+	root := cciptypes.Bytes32{}
 	commitReports := map[cciptypes.ChainSelector][]plugintypes.ExecutePluginCommitDataWithMessages{
 		1: {
 			{
@@ -740,18 +353,7 @@ func TestPlugin_Outcome_CommitReportsMergeError(t *testing.T) {
 	}
 
 	commitReports := map[cciptypes.ChainSelector][]plugintypes.ExecutePluginCommitDataWithMessages{
-		1: {
-			{
-				ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
-					MerkleRoot: mustParseByteStr(""),
-				},
-			},
-			{
-				ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
-					MerkleRoot: mustParseByteStr(""),
-				},
-			},
-		},
+		1: {},
 	}
 	observation, err := plugintypes.NewExecutePluginObservation(commitReports, nil).Encode()
 	require.NoError(t, err)
